# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是 AIA 社团论坛的后端服务，支持 Markdown 文章管理与静态资源分类存储。",
        "title": "AIA 社团博客 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/articles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "获取文章列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/article.ArticleListResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "发布文章",
                "parameters": [
                    {
                        "description": "文章数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/article.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "{\"id\": \"uuid\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/article.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/article.Article"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/article.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "编辑文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新的数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/article.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/article.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文章 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/assets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "获取静态资源列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "路径范围",
                        "name": "scope",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "article",
                            "shared"
                        ],
                        "type": "string",
                        "description": "类型",
                        "name": "kind",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assets"
                ],
                "summary": "上传静态资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "第一层路径",
                        "name": "scope",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件名主干",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "是否允许覆盖同路径资源",
                        "name": "overwrite",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/asset.AssetRecord"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "删除静态资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "资源完整路径 (如 /images/capoo.webp)",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "终极管理员登录 (仅依赖高熵动态口令)",
                "parameters": [
                    {
                        "description": "登录参数 (将生成的 16 位强动态口令填入 password 字段)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "article.Article": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "Author 作者姓名或昵称",
                    "type": "string",
                    "example": "负责人"
                },
                "category": {
                    "description": "Category 文章分类",
                    "type": "string",
                    "example": "技术分享"
                },
                "content": {
                    "description": "Content 文章正文",
                    "type": "string",
                    "example": "# 第一章\n内容如下..."
                },
                "cover": {
                    "description": "Cover 文章封面图 URL",
                    "type": "string",
                    "example": "https://example.com/cover.jpg"
                },
                "createdAt": {
                    "description": "CreatedAt 创建时间",
                    "type": "string"
                },
                "defaultMode": {
                    "description": "DefaultMode 展示模式：article, slide, homework",
                    "type": "string",
                    "example": "article"
                },
                "id": {
                    "description": "ID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "title": {
                    "description": "Title 文章标题",
                    "type": "string",
                    "example": "关于后端架构的深度思考"
                },
                "updatedAt": {
                    "description": "UpdatedAt 更新时间",
                    "type": "string"
                }
            }
        },
        "article.ArticleListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/article.Article"
                    }
                }
            }
        },
        "article.CreateRequest": {
            "type": "object",
            "required": [
                "author",
                "category",
                "content",
                "defaultMode",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string",
                    "example": "作者"
                },
                "category": {
                    "type": "string",
                    "example": "分类"
                },
                "content": {
                    "type": "string",
                    "example": "# 正文"
                },
                "cover": {
                    "type": "string"
                },
                "defaultMode": {
                    "type": "string",
                    "enum": [
                        "article",
                        "slide",
                        "homework"
                    ],
                    "example": "article"
                },
                "title": {
                    "type": "string",
                    "example": "标题"
                }
            }
        },
        "article.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "错误信息描述"
                }
            }
        },
        "asset.AssetRecord": {
            "type": "object",
            "properties": {
                "contentType": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "markdownValue": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "uploadedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "auth.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "auth.LoginRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "x6+xJ/c5fsWaVyxu"
                }
            }
        },
        "auth.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}